+++ db/upgrade.php
@@ -50,5 +50,14 @@ function xmldb_theme_boost_campus_upgrade($oldversion) {
         upgrade_plugin_savepoint(true, 2018051701, 'theme', 'boost_campus');
     }
 
+    if ($oldversion < 2018121700) {
+        // The setting "theme_boost_campus|incoursesettingsswitchtorole" has been renamed because the setting was
+        // upgraded with another option.
+        // Therefore set the old config to null.
+        set_config('incoursesettingsswitchtorole', null, 'theme_boost_campus');
+
+        upgrade_plugin_savepoint(true, 2018121700, 'theme', 'boost_campus');
+    }
+
     return true;
 }