+++ locallib.php
@@ -261,8 +261,11 @@ function theme_boost_campus_get_incourse_settings() {
             $node = $PAGE->settingsnav->find('courseadmin', navigation_node::TYPE_COURSE);
             // Check if $node is not empty for other pages like for example the langauge customization page.
             if (!empty($node)) {
-                // If the setting 'incoursesettingsswitchtorole' is enabled add these to the $node.
-                if (get_config('theme_boost_campus', 'incoursesettingsswitchtorole') == 'yes' && !is_role_switched($COURSE->id)) {
+                // If the setting 'incoursesettingsswitchtoroleposition' is set either set to the option 'yes'
+                // or to the option 'both', then add these to the $node.
+                if (((get_config('theme_boost_campus', 'incoursesettingsswitchtoroleposition') == 'yes') ||
+                    (get_config('theme_boost_campus', 'incoursesettingsswitchtoroleposition') == 'both'))
+                    && !is_role_switched($COURSE->id)) {
                     // Build switch role link
                     // We could only access the existing menu item by creating the user menu and traversing it.
                     // So we decided to create this node from scratch with the values copied from Moodle core.